refactor(test): P5-3 extract React-runtime parsers from test_design_import
Phase 5 P5-3 follow-up. React-runtime parser cluster extracted from
test_design_import.cpp into test_design_import_react_runtime.cpp
(21 TEST_CASEs, 183 assertions, 1,033 lines).

Covers TSX/runtime React parsers for every supported design-tool
source — shared contract (parse fixture, materialize via host React
shim, accept sanitized TSX, reject out-of-matrix surfaces):
  * parse_v0_tsx + parse_v0_dev_react           (v0.dev)
  * parse_figma_make_react                      (Figma Make)
  * parse_stitch_react                          (Stitch)
  * parse_react_native_export                   (React Native)
  * parse_pencil_react                          (Pencil)

The 175-line anonymous-namespace helpers block (read_fixture, asset_text,
minimal_host_react_dom_shim, etc.) is duplicated into the extracted TU
so it can stand alone (parent file keeps its own copy for the remaining
non-React tests).

test_design_import.cpp: 2,182 → 1,356 (-826 lines, well under the
< 2,000 P5-3 target).
